+++ Changelog.md
@@ -7,6 +7,8 @@ OpenCore Changelog
 - Fixed handling multinode device paths in LoadedImage and elsewhere
 - Changed OpenCanopy image directory to support directory prefixes
 - Changed OpenCanopy preferred image set to `Acidanthera\GoldenGate`
+- Added content flavour system allowing custom boot entry icons compatible across icon packs
+- Added automatic flavour detection for macOS boot entries
 
 #### v0.6.9
 - Fixed out-of-sync cursor movement rectangle when loading e.g. CrScreenshotDxe